Re-use file status when opening a new stream (#26)
`FileSystem.open` is pretty expensive for S3 without a file status, since it needs to get the HEAD info of the file. Even worse, in the async path we are calling `FileSystem.open` for every column read in a sequential manner.

This changes it to re-use the file status so the step can be skipped.