Add an option to return compressed pages (#50)
This adds an option: `parquet.read.compressed-pages.enabled`, to return compressed dictionary and data pages when enabled. By default it is false.

Added a new method `isCompressed` to Parquet's `Page` class. Note that `DataPageV2` already has the same method so this makes it override the basic one.

Added a basic unit test in `TestParquetFileReader`: it doesn't check the content of the pages, but just simply test whether the `isCompressed` is properly set according to the config.

This is useful when we want to (optionally) move decompression to the native side.